+++ public/directives/wz-logtest/components/logtest.tsx
@@ -49,7 +49,7 @@ export const Logtest = compose(
     setValue(e.target.value.split('\n').filter((item) => item));
   };
 
-  const formatResult = (result) => {
+  const formatResult = (result, alert) => {
     return (
       `**Phase 1: Completed pre-decoding. \n    ` +
       `full event:  ${result.full_log || '-'}  \n    ` +
@@ -58,10 +58,8 @@ export const Logtest = compose(
       `program_name: ${(result.predecoder || '').program_name || '-'} \n\n` +
       `**Phase 2: Completed decoding. \n    ` +
       `name: ${(result.decoder || '').name || '-'} \n    ` +
-      `parent: ${(result.decoder || '').parent || '-'} \n    ` +
-      `srcip: ${(result.data || '').srcip || '-'}  \n    ` +
-      `srcport: ${(result.data || '').srcport || '-'} \n    ` +
-      `srcuser: ${(result.data || '').srcuser || '-'} \n\n` +
+      `${(result.decoder || '').parent ? `parent: ${(result.decoder || '').parent} \n    ` : ''}` +
+      `data: ${JSON.stringify(result.data || '-', null, 6).replace('}', '    }')} \n\n` +
       `**Phase 3: Completed filtering (rules). \n    ` +
       `id: ${(result.rule || '').id || '-'} \n    ` +
       `level: ${(result.rule || '').level || '-'} \n    ` +
@@ -79,30 +77,38 @@ export const Logtest = compose(
       `nist_800_53: ${JSON.stringify((result.rule || '').nist_800_53 || '-')} \n    ` +
       `pci_dss: ${JSON.stringify((result.rule || '').pci_dss || '-')} \n    ` +
       `tsc: ${JSON.stringify((result.rule || '').tsc || '-')} \n` +
-      `**Alert to be generated. \n\n\n`
+      `${alert ? `**Alert to be generated. \n\n\n` : '\n\n'}`
     );
   };
 
   const runAllTests = async () => {
     setTestResult('');
     setTesting(true);
-    try{
-      const responsesLogtest = await Promise.all(value.map(async event => {
-        const body = {
-          log_format: 'syslog',
-          location: 'logtest',
-          event: event,
-        };
-        return await WzRequest.apiReq('PUT', '/logtest', body);
-      }));
-      const testResults = responsesLogtest.map(response => 
-        response.data.data.alert
-          ? formatResult(response.data.data.output)
+    try {
+      const responsesLogtest = await Promise.all(
+        value.map(async (event) => {
+          const body = {
+            log_format: 'syslog',
+            location: 'logtest',
+            event: event,
+          };
+          return await WzRequest.apiReq('PUT', '/logtest', body);
+        })
+      );
+      const testResults = responsesLogtest.map((response) =>
+        response.data.data.output.rule || ''
+          ? formatResult(response.data.data.output, response.data.data.alert)
           : `No result found for:  ${response.data.data.output.full_log} \n\n\n`
       );
       setTestResult(testResults);
-    }finally{
+    } finally {
       setTesting(false);
+    }
+  };
+
+  const handleKeyPress = async (event) => {
+    if (event.ctrlKey && event.key === 'Enter') {
+      await runAllTests();
     }
   };
 
@@ -115,6 +121,7 @@ export const Logtest = compose(
           aria-label=""
           rows={props.showClose ? 10 : 4}
           onChange={onChange}
+          onKeyPress={handleKeyPress}
         />
         <EuiSpacer size="m" />
         <EuiFlexItem grow={false}>