+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 ## Wazuh v4.2.0 - Kibana 7.10.0 , 7.10.2 - Revision 4202
 
-### Added
+### Changed
 - Remove module titles [#3160](https://github.com/wazuh/wazuh-kibana-app/pull/3160)
 
 ## Wazuh v4.2.0 - Kibana 7.10.0 , 7.10.2 - Revision 4201