+++ CHANGELOG.md
@@ -41,6 +41,10 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Fixed RBAC issue with agent group permissions [#3181](https://github.com/wazuh/wazuh-kibana-app/pull/3181)
 - Fixed change index pattern from menu doesn't work [#3187](https://github.com/wazuh/wazuh-kibana-app/pull/3187)
 
+## Wazuh v4.1.5 - Kibana 7.10.0 , 7.10.2 - Revision 4106
+
+- Adapt for Wazuh 4.1.5
+
 ## Wazuh v4.1.4 - Kibana 7.10.0 , 7.10.2 - Revision 4105
 
 - Adapt for Wazuh 4.1.4